Adding further features, adding encoding

+++ src/benchmark/cost_model_calibration/configuration/calibration_configuration.hpp
@@ -20,7 +20,7 @@ namespace opossum {
 
         std::vector<CalibrationTableSpecification> table_specifications;
         std::string output_path;
-        int calibration_runs;
+        size_t calibration_runs;
 
     };
 }  // namespace opossum
